requirements.txt
pyfluidsynth==1.3.4
python-rtmidi==1.5.8
keyboard==0.13.5
mido==1.3.3